move manager.go to controller package
