+++ pkg/controller/manager.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package manager
+package controller
 
 import (
 	"k8s.io/client-go/rest"
