+++ pkg/controller/manager_test.go
@@ -14,6 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package manager
+package controller
 
 // (khos2ow): TODO
